pyproject.toml
[tool.poetry]
name = "mlops_assignment"
version = "0.2.0"
description = ""
authors = ["amber <amberyeo21@gmail.com>", "javerine <javerinetan05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "2.1.4"
flask = "3.0.3"
hydra-core = "^1.3.2"
dvc = "^3.53.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
